Move some types from `tasks` around to better fits
